requirements.txt
langchain==0.3.23
langchain-community==0.3.21
langchain-ollama==0.3.2
chromadb==1.0.4
PyYAML==6.0.2
python-docx==1.1.2
langchain-chroma==0.2.3
great-expectations==0.18.19